heic2jpg/requirements.txt
Pillow
pillow-heif
watchdog